2.0.8RC2 fixes PR1283148 and 1281189.

+++ tests.php4/session.php
@@ -12,6 +12,7 @@ if($session->isNew()) {
   echo "new session\n";
   $session->put("a", 1);
   $session->put("b", 5);
+  $session->put("c", null);
 }
 else {
   echo "cont session\n";
@@ -19,8 +20,10 @@ else {
 
 $session->put("a", $session->get("a")+1);
 $session->put("b", $session->get("b")-1);
-     
+
 $val=$session->get("a");
+$c=$session->get("c");
+if($c!=null) {echo "test failed"; exit(1);}
 echo "session var: $val\n";
 
 if($session->get("b")==0) $session->destroy();
